ingest_mongo/requirements.txt
python-dotenv
boto3
pymongo
